+++ CameraPlus/Behaviours/CameraPreviewQuad.cs
@@ -11,6 +11,26 @@ namespace CameraPlus.Behaviours
 		private GameObject _cameraQuad;
 
 		internal Material _previewMaterial = new Material(Plugin.cameraController.Shaders["BeatSaber/BlitCopyWithDepth"]);
+		internal Material _cubeMaterial = new Material(Plugin.cameraController.Shaders["BeatSaber/BlitCopyWithDepth"]);
+
+		public bool IsDisplayMaterialVROnly
+		{
+			get{
+				return _previewMaterial.GetFloat("_IsVRCameraOnly") == 1;
+			}
+			set{
+				if (value)
+				{
+					_previewMaterial.SetFloat("_IsVRCameraOnly", 1);
+					_cubeMaterial.SetFloat("_IsVRCameraOnly", 1);
+				}
+				else
+				{
+					_previewMaterial.SetFloat("_IsVRCameraOnly", 0);
+					_cubeMaterial.SetFloat("_IsVRCameraOnly", 0);
+				}
+			}
+		}
 
 		public void Awake()
 		{
@@ -21,9 +41,10 @@ namespace CameraPlus.Behaviours
 			cam = camera;
 			_cameraCube = GameObject.CreatePrimitive(PrimitiveType.Cube);
 			DontDestroyOnLoad(_cameraCube);
-			_cameraCube.transform.localScale = new Vector3(0.15f, 0.15f, 0.22f);
+			_cameraCube.GetComponent<MeshRenderer>().material = _cubeMaterial;
+			SetCameraCubeSize(PluginConfig.Instance.CameraCubeSize);
+			Logger.log.Notice($"Camera Aspect {cam._cam.aspect}");
 			_cameraCube.name = "CameraCube";
-			_cameraCube.layer = PluginConfig.Instance.CameraQuadLayer;
 			_cameraCube.transform.SetParent(transform);
 			_cameraCube.transform.localPosition = Vector3.zero;
 			_cameraCube.transform.localEulerAngles = Vector3.zero;
@@ -32,17 +53,32 @@ namespace CameraPlus.Behaviours
 			DontDestroyOnLoad(_cameraQuad);
 			DestroyImmediate(_cameraQuad.GetComponent<Collider>());
 			_cameraQuad.GetComponent<MeshRenderer>().material = _previewMaterial;
-			_cameraQuad.transform.SetParent(_cameraCube.transform);
-			_cameraQuad.transform.localPosition = new Vector3(-1f * ((cam._cam.aspect - 1) / 2 + 1), 0, 0.22f);
-			_cameraQuad.transform.localEulerAngles = new Vector3(0, 180, 0);
-			_cameraQuad.transform.localScale = new Vector3(cam._cam.aspect, 1, 1);
-			_cameraQuad.layer = PluginConfig.Instance.CameraQuadLayer;
+			_cameraQuad.transform.SetParent(transform);
+			SetCameraQuadPosition(PluginConfig.Instance.CameraQuadPosition, PluginConfig.Instance.CameraCubeSize);
+			IsDisplayMaterialVROnly = cam.Config.PreviewCameraVROnly;
 		}
 		public void OnPointerClick(PointerEventData eventData)
 		{
 			if (!(eventData.currentInputModule is VRUIControls.VRInputModule)) return;
 			if (cam.Config.cameraLock.lockCamera) return;
 			CameraMoverPointer.BeginDragCamera(cam);
+		}
+		public void SetCameraQuadPosition(string quadPosition, float cubeScale = 1)
+        {
+			_cameraQuad.transform.localEulerAngles = new Vector3(0, 180, 0);
+			_cameraQuad.transform.localScale = new Vector3(0.15f * (PluginConfig.Instance.CameraQuadStretch ? 1.7778f : cam._cam.aspect), 0.15f * 1, 1); //1.7778f = 16 / 9
+			if (quadPosition == "Top")
+				_cameraQuad.transform.localPosition = new Vector3(0, _cameraQuad.transform.localScale.y / 2 + _cameraCube.transform.localScale.y / 2, 0.05f * cubeScale);
+			else if (quadPosition == "Bottom")
+				_cameraQuad.transform.localPosition = new Vector3(0, -_cameraQuad.transform.localScale.y / 2 - _cameraCube.transform.localScale.y / 2, 0.05f * cubeScale);
+			else if (quadPosition == "Left")
+				_cameraQuad.transform.localPosition = new Vector3(_cameraQuad.transform.localScale.x / 2 + _cameraCube.transform.localScale.x / 2, 0, 0.05f * cubeScale);
+			else
+				_cameraQuad.transform.localPosition = new Vector3(-_cameraQuad.transform.localScale.x / 2 - _cameraCube.transform.localScale.x / 2, 0, 0.05f * cubeScale);
+		}
+		public void SetCameraCubeSize(float cubeScale)
+        {
+			_cameraCube.transform.localScale = new Vector3(0.15f * cubeScale, 0.15f * cubeScale, 0.22f * cubeScale);
 		}
 	}
 }